requirements.txt
beautifulsoup4==4.13.3
openai==1.64.0
googlesearch-python==1.3.0
python-dotenv==1.0.1
rich==13.9.4